*.[ch]: make a distinction between SCM as a generic name for a Scheme object (now a void*), and SCM as 32 bit word for storing tags and immediates (now a long int).  Introduced SCM_ASWORD and SCM_ASSCM for conversion. Fixed various dubious code in the process: arbiter.c (use macros), unif.c (scm_array_p),

+++ libguile/feature.c
@@ -142,7 +142,7 @@ print_hook (SCM hook, SCM port, scm_print_state *pstate)
     }
   scm_intprint (SCM_HOOK_ARITY (hook), 10, port);
   scm_putc (' ', port);
-  scm_intprint (hook, 16, port);
+  scm_intprint ((int)hook, 16, port);
   ls = SCM_HOOK_PROCEDURES (hook);
   while (SCM_NIMP (ls))
     {
